requirement.txt
selenium==3.141.0
numpy
opencv-python
pillow
tqdm
